refactor(web): extract GenUI Possibilities feature boundary
