+++ examples/web/genui-possibilities.html
@@ -100,6 +100,6 @@
     </main>
 
     <div id="toast" class="toast" role="status" aria-live="polite"></div>
-    <script type="module" src="/src/genui-possibilities.js"></script>
+    <script type="module" src="/src/entries/genui-possibilities.js"></script>
   </body>
 </html>
