+++ examples/web/src/entries/genui-possibilities.js
@@ -1,0 +1,3 @@
+import { mountGenuiPossibilities } from '../features/genui-possibilities/browser/mount.js';
+
+mountGenuiPossibilities();
