+++ examples/web/src/features/genui-possibilities/core/journey-state.js
